requirements.txt
streamlit
pandas
websocket-client
plotly